requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
distlib
dj-database-url
Django
django-allauth
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
environ
filelock
idna
Markdown
MouseInfo
oauthlib
pillow
pipdeptree
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python-decouple
python-dotenv
python3-openid
pytweening
pytz
requests
requests-oauthlib
sqlparse
typing_extensions
tzdata
urllib3
virtualenv